+++ src/test/integration/runTest.ts
@@ -1,4 +1,3 @@
-// @ts-nocheck
 import * as path from "path";
 import { runTests } from "@vscode/test-electron";